requirements.txt
fastapi
uvicorn
grpcio
pandas